fix(server): format size counts portably (#23)
## Summary

- replace server `size_t` `%zu` formatting with the existing `PRIuMAX`
plus `uintmax_t` convention
- cover the metrics, pathfinder, and waypoint sites checked by MinGW
printf attributes
- retain exact decimal output semantics and keep the configured Windows
stdio mode unchanged
- keep metrics test identifier formatting portable

Fixes the Windows server package compile failure exposed by rehearsal
run 31242894929, job 93066714965.

## Validation

- `./atrinik profile show classic-windows-server-metrics-format`
- `./atrinik build server --profile
classic-windows-server-metrics-format --test` (34/34 passed with
`ENABLE_WARNING_ERRORS=ON`)
- exact digest-pinned Windows release container: `bash
tools/build-windows-package.sh build/packages`
- generated `atrinik-classic-server-0.0.0-windows-x86_64.zip` passed
`unzip -tq`
- `python3 server/tools/dependencies.py validate`
- `python3 server/tools/dependencies.py verify`
- `python3 tools/verify_import_history.py`
- `clang-format --dry-run --Werror` on all changed C files
- `git diff --check`

+++ server/src/server/pathfinder.c
@@ -207,9 +207,9 @@ static bool clioptions_option_pathfinder_max_nodes(const char *arg, char **errms
     }
 
     LOG(INFO,
-        "Pathfinding generated-state budget changed from %zu to %zu",
-        path_max_generated,
-        (size_t)value);
+        "Pathfinding generated-state budget changed from %" PRIuMAX " to %" PRIuMAX,
+        (uintmax_t)path_max_generated,
+        (uintmax_t)value);
     path_max_generated = (size_t)value;
     return true;
 }
